requirements.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
PyYAML==5.1.2
requests==2.22.0
robotframework==3.1.2
robotframework-seleniumlibrary==4.1.0
selenium==3.141.0
urllib3==1.25.6